+++ src/index.js
@@ -2,8 +2,17 @@
 
 export { default as FormComponent } from './FormComponent';
 export { default as useForm } from './use-form';
+export { default as useFormGroup } from './use-form-group';
+export { default as useFormGroupItem } from './use-form-group-item';
 export { default as withForm } from './with-form';
-export type { $Optional, FieldProp, FormErrors, FormObject } from './types';
+export type {
+  $Optional,
+  FieldProp,
+  FormErrors,
+  FormGroup,
+  FormGroupItemControls,
+  FormObject,
+} from './types';
 
 export { UntypedFormComponent } from './untyped-exports';
 export type { UntypedField, UntypedFormObject } from './untyped-exports';